inheritance/requirements.txt
django-mptt
django-dirtyfields